requirements.txt
tensorflow==0.11.0rc2
PyYAML==3.11
boto3==1.4.1
argparse=1.4.0